Handling an error caused by Evennia's spawn system.
Evennia does not have a proper place for a post-spawn hook.
Now ObjectParent.at_post_spawn() will be called when using
* my util.spawn.spawn() or
* the spawn command, overridden from my commands.CustomSpawn

The overridden spawn command is a brute force copy of the original command
and may need to be updated if that function is updated upstream.

My spawn function is a simple wrapper around
evennia.prototypes.spawner.spawn()
and should be preferred over that in all cases.