Stances and Stance Management
Initial implementation of skillsets with stance, job, and role variants.  Creates verbs to handle them.  These map aspects to skills with a limited number of slots, and are intended to made skill use less annoying as you level up.

Adds variant Create command that attempts to put created items in your hand if at all possible, rather than putting them in phantom, inaccessible evennia inventory.

Adds initial support for room event tags to scene verb, but does not implement scene objects having multiple descriptions to account for them.  Adds display_section to objects to better handle scene objects, and allows exits to be scene objects.  Adds the ability for objects to be in the display footer of a room, but that is also below the item/character lists, so this is best for event stuff.

Adds some name hints and footer information for exits, so that if an exit has a cardinal direction alias but isn't named that, that will be obvious.

Expands AttributeProperty so that attribute category can be gotten dynamically.  This was used for Skillsets, since stances/jobs/roles are stored in their respective attribute categories, but with the same underlying code.