Skills, scenes, xp, and some cleaning
SceneObjects now support limited variable replacement during an event.  This is not comprehensive, but for example, some objects may become invisible during an event, or invisible except during an event.  This is useful for some scene elements to just "appear' when an event is triggered.
The Scene verb can now create SceneObjects, set scene-relevant variables, and set the event-variable replacements.  SceneObjects have a helper function to get the current state of replaceable variables, although this could be cleaner.

The Stat and Skill verbs are now using custom EvTables, each much better in different ways.  Both also have a /raw subcommand for debug examination of data values.  The Skill display table is highly extensible without losing detail, and now will even display if the skill is currently assigned to a Stance/Job/Role (and which).  The Stat display can focus on any stat, row, or column, and now displays per-stat xp.

The whole thing with manually linking skills to stats is gone.  Official stat linkages are determined by the skill TOML database, which needs to be updated as skills are created.  Skills will always have a (limited) stat bonus.  The rolling mechanisms still need to be updated to reflect the skillset restrictions.  Skillsets now have hooks to help manage skills being added and removed, activated and deactivated.

Minor:
* Moved skillset commands to work as aliases of a single version.
* Skill streak cooldowns are now working, my previous version might have worked if I was just not an idiot.
* Stat XP is refactored a bit, though it still needs more work.
* Bugfixen